feat: /ban 实现应该也是在userInfoService内比较好，因为是管理员根据用户来ban用户，和最初登录关联不大
